# gh-117686: Improve the performance of ntpath.expanduser()