--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-09-18-13-08.gh-issue-117686.4oIvxF.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-09-18-13-08.gh-issue-117686.4oIvxF.rst
@@ -1,0 +1,1 @@
+Speedup :func:`os.path.expanduser()` on Windows for usernames.