### Benchmark

```bat
::test.bat
@echo off
echo 1 character && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~a')" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~a')"
echo 10 characters && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~' + 'a' * 10)" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~' + 'a' * 10)"
echo 100 characters && python -m timeit -s "import before.ntpath" "before.ntpath.expanduser('~' + 'a' * 100)" && python -m timeit -s "import after.ntpath" "after.ntpath.expanduser('~' + 'a' * 100)"
```

```none
1 character
50000 loops, best of 5: 5.93 usec per loop # before
50000 loops, best of 5: 5.77 usec per loop # after
# -> 1.03x faster
10 characters
50000 loops, best of 5: 7.49 usec per loop # before
50000 loops, best of 5: 6.43 usec per loop # after
# -> 1.16x faster
100 characters
20000 loops, best of 5: 19.5 usec per loop # before
20000 loops, best of 5: 12.2 usec per loop # after
# -> 1.60x faster
```

<!-- gh-issue-number: gh-117686 -->
* Issue: gh-117686
<!-- /gh-issue-number -->